+++ src/components/forms/PasswordResetForm.vue
@@ -1,0 +1,64 @@
+<template>
+  <form class="space-y-6" @submit.prevent="">
+    <div>
+      <label for="email" class="block pl-3 text-lg/6 font-medium text-white"
+        >이메일</label
+      >
+
+      <BaseInput></BaseInput>
+    </div>
+
+    <div>
+      <div class="flex items-center justify-between px-3">
+        <label
+          for="password"
+          class="block text-lg/6 font-medium text-white"
+          placeholder="1234"
+          >비밀번호</label
+        >
+        <div class="text-sm">
+          <a
+            href="#"
+            class="font-semibold text-white hover:text-gray-500"
+            @click="goFind"
+            >기억이 안나십니까?</a
+          >
+        </div>
+      </div>
+
+      <BaseInput></BaseInput>
+    </div>
+
+    <div class="px-3 py-3">
+      <button
+        type="submit"
+        class="neon-border flex w-full justify-center rounded-md px-3 py-3 text-lg font-extrabold text-white shadow-xs"
+        :style="`--neon-color:${neonColor}`"
+      >
+        로그인
+      </button>
+    </div>
+  </form>
+
+  <p
+    class="mt-5 overflow-hidden text-center text-lg/5 text-ellipsis whitespace-nowrap text-white"
+  >
+    아직도 회원이 아니신가요?
+    <a
+      href="#"
+      class="font-semibold overflow-ellipsis text-white hover:text-indigo-500"
+      @click="goSign"
+      >회원가입하기</a
+    >
+  </p>
+</template>
+
+<script setup>
+import BaseInput from "../BaseInput.vue";
+
+const neonColor = "#00f0ff";
+</script>
+
+<style scoped>
+@import "../../styles/neon.css";
+</style>
